requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.9.5=py310hb372a2b_0
aiosignal=1.3.1=pyhd8ed1ab_0
artiq=8.8847+a68846d=0
artiq-comtools=1.3=0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
brotli-python=1.1.0=py310h9e9d8ca_1
bzip2=1.0.8=h10d778d_5
c-ares=1.28.1=h10d778d_0
ca-certificates=2024.6.2=h8857fd0_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2024.6.2=pyhd8ed1ab_0
cffi=1.16.0=py310hdca579f_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
frozenlist=1.4.1=py310hb372a2b_0
gettext=0.22.5=h5ff76d1_2
gettext-tools=0.22.5=h5ff76d1_2
glib=2.80.2=h736d271_1
glib-tools=2.80.2=h959cb30_1
gmqtt=0.6.16=pypi_0
gst-plugins-base=1.24.4=ha0a8333_0
gstreamer=1.24.4=hbc75551_0
h5py=3.9.0=nompi_py310h164cd36_101
hdf5=1.14.1=nompi_hedada53_100
icu=73.2=hf5e326d_0
idna=3.7=pyhd8ed1ab_0
krb5=1.21.2=hb884880_0
levenshtein=0.25.1=py310h5daac23_1
libaec=1.1.3=h73e2aa4_0
libasprintf=0.22.5=h5ff76d1_2
libasprintf-devel=0.22.5=h5ff76d1_2
libblas=3.9.0=17_osx64_openblas
libcblas=3.9.0=17_osx64_openblas
libclang-cpp15=15.0.7=default_h7151d67_5
libclang13=18.1.7=default_h0449792_0
libcurl=8.8.0=hf9fcc65_0
libcxx=17.0.6=h88467a6_0
libedit=3.1.20191231=h0678c8f_2
libev=4.33=h10d778d_2
libffi=3.4.2=h0d85af4_5
libgettextpo=0.22.5=h5ff76d1_2
libgettextpo-devel=0.22.5=h5ff76d1_2
libgfortran=5.0.0=11_3_0_h97931a8_32
libgfortran5=12.2.0=he409387_32
libgit2=1.8.1=h59467ec_1
libglib=2.80.2=h736d271_1
libiconv=1.17=hd75f5a5_2
libintl=0.22.5=h5ff76d1_2
libintl-devel=0.22.5=h5ff76d1_2
libjpeg-turbo=3.0.0=h0dc2134_1
liblapack=3.9.0=17_osx64_openblas
libllvm11=11.1.0=h8fb7429_5
libllvm14=14.0.6=hc8e404f_4
libllvm15=15.0.7=hbedff68_4
libllvm18=18.1.7=h9ce406d_1
libnghttp2=1.58.0=h64cf6d3_1
libogg=1.3.4=h35c211d_1
libopenblas=0.3.23=openmp_h429af6e_0
libopus=1.3.1=hc929b4f_1
libpng=1.6.43=h92b6c6a_0
libpq=16.3=h4501773_0
libsqlite=3.46.0=h1b8f9f3_0
libssh2=1.11.0=hd019ec5_0
libvorbis=1.3.7=h046ec9c_0
libxml2=2.12.7=h3e169fe_1
libzlib=1.3.1=h87427d6_1
lld=18.1.7=hd78e567_0
llvm-openmp=18.1.7=h15ab845_0
llvm-tools=18.1.7=h9ce406d_1
llvmlite=0.43.0=py310h515a046_0
multidict=6.0.5=py310h90acd4f_0
mysql-common=8.3.0=hfd7a639_4
mysql-libs=8.3.0=ha9146f8_4
ncurses=6.5=h5846eda_0
nspr=4.35=hea0b92c_0
nss=3.101=he7eb89d_0
numpy=1.26.4=py310h4bfa8fc_0
openssl=3.3.1=h87427d6_0
packaging=24.1=pyhd8ed1ab_0
pcre2=10.44=h7634a1b_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.2=pyhd8ed1ab_0
ply=3.11=pyhd8ed1ab_2
pooch=1.8.2=pyhd8ed1ab_0
prettytable=3.10.0=pyhd8ed1ab_0
pybase64=1.3.2=py310hb372a2b_1
pycparser=2.22=pyhd8ed1ab_0
pygit2=1.15.0=py310h936d840_0
pyqt=5.15.9=py310hecc045f_5
pyqt5-sip=12.12.2=py310h018f80b_5
pyqtgraph=0.13.7=pyhd8ed1ab_0
pyserial=3.5=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.14=h00d2728_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-levenshtein=0.25.1=pyhd8ed1ab_0
python-lmdb=1.4.1=py310h0c761eb_1
python_abi=3.10=4_cp310
pythonparser=1.4=0
qasync=0.27.1=pyh9208f05_0
qt-main=5.15.8=hecaf5c3_21
rapidfuzz=3.9.3=py310he0a0c5d_0
readline=8.2=h9e318b2_1
regex=2024.5.15=py310h936d840_0
requests=2.32.3=pyhd8ed1ab_0
scipy=1.11.1=py310h3900cf1_0
setuptools=70.0.0=pyhd8ed1ab_0
sip=6.7.12=py310had63691_0
sipyco=1.7=0
six=1.16.0=pyh6c4a22f_0
tk=8.6.13=h1abcd95_1
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
tqdm=4.66.4=pyhd8ed1ab_0
typing-extensions=4.12.2=hd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h0c530f3_0
urllib3=2.2.2=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_1
windfreak=0.3.0=pyhd8ed1ab_0
xz=5.2.6=h775f41a_0
yarl=1.9.4=py310hb372a2b_0
zstd=1.5.6=h915ae27_0
